refactor: restructure session strategies
- Move strategies under business/service/
- Add express and fastify helper classes
- Export node-session-util from package
- Add contract tests and vitest config
- Remove IDE-generated iml file
