+++ package.json
@@ -25,6 +25,10 @@
 			"types": "./dist/index.d.ts",
 			"import": "./dist/index.js"
 		},
+		"./node-session-util": {
+			"types": "./dist/util/node-session-util.d.ts",
+			"import": "./dist/util/node-session-util.js"
+		},
 		"./*": {
 			"types": "./dist/*.d.ts",
 			"import": "./dist/*.js"
@@ -69,6 +73,7 @@
 		"test-coverage-unit-publish": "codecov",
 		"test:e2e": "cd test && npm run test",
 		"test:int": "vitest --config=./vitest.config.int.ts",
+		"test:contract": "vitest --config=./vitest.config.contract.ts",
 		"test:unit": "vitest --config=./vitest.config.unit.ts",
 		"tsc": "tsc -p tsconfig.build.json",
 		"tsc-check": "npm run tsc -- --noEmit",
@@ -94,6 +99,7 @@
 		"fastify-plugin": "^6.0.0"
 	},
 	"devDependencies": {
+		"@beecode/msh-test-contractor": "^0.4.2",
 		"@commitlint/cli": "^21.0.2",
 		"@commitlint/config-conventional": "^21.0.2",
 		"@commitlint/prompt": "^21.0.2",
@@ -113,6 +119,7 @@
 		"typedoc": "^0.28.19",
 		"typedoc-plugin-markdown": "^4.12.0",
 		"typescript": "^6.0.3",
+		"vite-tsconfig-paths": "^6.1.1",
 		"vitest": "^4.1.8"
 	},
 	"engines": {
