+++ src/business/service/session-strategy/fast-als-strategy.ts
@@ -2,7 +2,7 @@
 // @ts-expect-error
 import fastAls from 'fast-als'
 
-import { type CreateSessionOptions, type SessionStrategy } from '#src/session-strategy/session-strategy.js'
+import { type CreateSessionOptions, type SessionStrategy } from '#src/business/service/session-strategy.js'
 
 export class FastAlsStrategy implements SessionStrategy {
 	clear(key: string): void {
