+++ vitest.config.unit.ts
@@ -1,5 +1,6 @@
-import baseDefaultConfig from './vitest.config.js'
 import { defineConfig, mergeConfig } from 'vite'
+
+import baseDefaultConfig from './vitest.config.js'
 
 export default mergeConfig(
 	baseDefaultConfig,
